requirements_standalone.txt
requests>=2.31.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.105.0
python-dateutil>=2.8.2
